+++ CHANGELOG.md
@@ -6,6 +6,10 @@
 
 * [#5976](https://github.com/rubocop-hq/rubocop/issues/5976): Warn for Rails Cops. ([@koic][])
 
+### Bug fixes
+
+* [#6936](https://github.com/rubocop-hq/rubocop/issues/6936): Fix `Layout/MultilineMethodArgumentLineBreaks` when bracket hash assignment on multiple lines. ([@maxh][])
+
 ## 0.70.0 (2019-05-21)
 
 ### New features